# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Career Connect API",
    "description": "Server for Career Connect frontend",
    "version": "1.0.0"
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Email": {
        "type": "string",
        "description": "User's email address",
        "example": "betty@email.com"
      },
      "Password": {
        "type": "string",
        "description": "User's password",
        "example": "cardigan"
      },
      "UserName": {
        "type": "string",
        "description": "User's name",
        "example": "Betty"
      },
      "UserId": {
        "type": "number",
        "description": "ID of the user",
        "example": 61021
      },
      "UserImage": {
        "type": "string",
        "description": "User's profile image (stored as base64). You will get out what you put in",
        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
      },
...
      "UserFull": {
        "type": "object",
        "description": "Detailed information of the user",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UserId"
          },
          "email": {
            "$ref": "#/components/schemas/Email"
          },
          "name": {
            "$ref": "#/components/schemas/UserName"
          },
          "image": {
            "$ref": "#/components/schemas/UserImage"
          },
          "watcheeUserIds": {
            "$ref": "#/components/schemas/UserIds"
          },
          "jobs": {
            "type": "array",
            "description": "Array of job postings by this user",
            "items": {
              "$ref": "#/components/schemas/JobFull"
            }
          }
        }
      },
      "TurnOn": {
        "type": "boolean",
        "description": "True if should add/enable, false if should remove/disable",
        "example": true
      },
...
      "UserIds": {
        "type": "array",
        "description": "List of user IDs",
        "items": {
          "$ref": "#/components/schemas/UserId"
        }
      },
      "JobFull": {
        "type": "object",
        "description": "Full details of the job posting",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/JobId"
          },
          "creatorId": {
            "$ref": "#/components/schemas/UserId"
          },
          "title": {
            "$ref": "#/components/schemas/JobTitle"
          },
          "image": {
            "$ref": "#/components/schemas/JobImage"
          },
          "start": {
            "$ref": "#/components/schemas/JobStart"
          },
          "description": {
            "$ref": "#/components/schemas/JobDescription"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "likes": {
            "type": "array",
            "description": "Array of all the users who have liked this job post",
            "items": {
              "$ref": "#/components/schemas/JobUserLike"
            }
          },
          "comments": {
            "type": "array",
            "description": "Array of all the comments users have made on this job post",
            "items": {
              "$ref": "#/components/schemas/JobUserComment"
            }
          }
        }
      },
...
      "JobTitle": {
        "type": "string",
        "description": "Name of the job post",
        "example": "COO for cupcake factory"
      },
      "JobImage": {
        "type": "string",
        "description": "Job's content image (stored as base64). You will get out what you put in",
        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
      },
      "JobStart": {
        "type": "string",
        "description": "Timestamp of when the job starts in simplified extended ISO format",
        "example": "2011-10-05T14:48:00.000Z"
      },
      "JobDescription": {
        "type": "string",
        "description": "Description of the job posting",
        "example": "Dedicated technical wizard with a passion and interest in human relationships"
      },
      "JobComment": {
        "type": "string",
        "description": "Comment on a job posting",
        "example": "This is a great opportunity, my email is hello@gmail.edu.au"
      },
      "JobUserLike": {
        "type": "object",
        "description": "Details about a particular like on a job posting",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/UserId"
          },
          "userEmail": {
            "$ref": "#/components/schemas/Email"
          },
          "userName": {
            "$ref": "#/components/schemas/UserName"
          }
        }
      },
      "JobUserComment": {
        "type": "object",
        "description": "Details about a particular comment on a job posting",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/UserId"
          },
          "userEmail": {
            "$ref": "#/components/schemas/Email"
          },
          "userName": {
            "$ref": "#/components/schemas/UserName"
          },
          "comment": {
            "$ref": "#/components/schemas/JobComment"
          }
        }
      },
      "JobId": {
        "type": "number",
        "description": "ID of the job",
        "example": 528491
      },
      "Timestamp": {
        "type": "string",
        "description": "Timestamp string in simplified extended ISO format",
        "example": "2011-10-05T14:48:00.000Z"
      }
    },
    "parameters": {
...
      "UserIdParam": {
        "name": "userId",
        "in": "query",
        "description": "User ID",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/UserId"
        }
      },
...
    },
    "responses": {
      "Success": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "AccessError": {
        "description": "Access error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid token"
                }
              }
            }
          }
        }
      },
      "InputError": {
        "description": "Input error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid input"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/user": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get the details of a specific user",
        "description": "Returns basic details of a specific user",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFull"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/AccessError"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update user's own profile",
        "description": "Updates the users' email, password, name, and/or image with the given details. It will only update defined fields provided. Any fields given that are undefined will remain unchanged.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "$ref": "#/components/schemas/Email"
                  },
                  "password": {
                    "$ref": "#/components/schemas/Password"
                  },
                  "name": {
                    "$ref": "#/components/schemas/UserName"
                  },
                  "image": {
                    "$ref": "#/components/schemas/UserImage"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/InputError"
          },
          "403": {
            "$ref": "#/components/responses/AccessError"
          }
        }
      }
    },
    "/user/watch": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Watching or unwatching another user",
        "description": "Specify whether this user should be watching or unwatching another user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "$ref": "#/components/parameters/UserEmail"
                  },
                  "turnon": {
                    "$ref": "#/components/schemas/TurnOn"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/InputError"
          },
          "403": {
            "$ref": "#/components/responses/AccessError"
          }
        }
      }
    }
  }
}